:recycle: refactor: 대기열 리팩토링
* 가게별 대기열 순번 조회 score 계산에서 rank로 변경
* 가게별 대기열 순번 조회 반환 타입 변경 Float -> Integer
* 대기열 조회 반환 타입 변경 Float -> Integer

+++ queue/src/main/java/com/quit/queue/application/service/dto/res/QueueResponse.java
@@ -1,6 +1,9 @@
 package com.quit.queue.application.service.dto.res;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +21,7 @@ public class QueueResponse {
     @AllArgsConstructor
     public static class UserScore {
         private Long userId;
-        private Float score;
+        private Integer score;
     }
 
     public QueueResponse(UUID storeId) {
@@ -26,7 +29,7 @@ public class QueueResponse {
         this.userScores = new ArrayList<>();
     }
 
-    public void addUserScore(Long userId, Float score) {
+    public void addUserScore(Long userId, Integer score) {
         if (this.userScores == null) {
             this.userScores = new ArrayList<>();
         }
